+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## v0.0.2 (2019-10-02)
+
+ - Add `get members` command
+ - Add `get members -o [normal, json, wide]` command
+
 ## v0.0.1 (2019-09-30)
 
  - Add `get repos` command